rework to move everything into new folders
