+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/DarkTheme">
-        <activity android:name=".Activities.LoginActivity">
+        <activity android:name=".Features.Login.Views.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
