+++ app/src/main/java/mtaubert/loginapplication/Data/DB/UserRoomDatabase.kt
@@ -1,10 +1,10 @@
-package mtaubert.loginapplication.RoomDatabase
+package mtaubert.loginapplication.Data.DB
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 
-import mtaubert.loginapplication.Data.User
-import mtaubert.loginapplication.DataAccessObjects.UserDAO
+import mtaubert.loginapplication.Data.DB.Model.User
+import mtaubert.loginapplication.Data.DB.DAO.UserDAO
 
 @Database(entities = [User::class], version = 1)
 abstract class UserRoomDatabase : RoomDatabase() {
