+++ app/src/main/java/mtaubert/loginapplication/Utils/Activities/BaseActivity.kt
@@ -1,4 +1,4 @@
-package mtaubert.loginapplication.Activities
+package mtaubert.loginapplication.Utils.Activities
 
 import android.content.Context
 import android.content.SharedPreferences
