+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activities.LoginActivity"
+        tools:context=".Features.Login.Views.LoginActivity"
         android:id="@+id/Background">
 
     <fragment
