+++ app/src/main/res/layout/fragment_account.xml
@@ -7,7 +7,7 @@
             android:id="@+id/constrainLayout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".Fragments.AccountFragment"
+            tools:context=".Features.Login.Views.AccountFragment"
             android:background="?attr/backgroundColor">
 
         <Button
