+++ app/src/main/res/layout/fragment_account_details.xml
@@ -7,7 +7,7 @@
             android:id="@+id/constraintLayout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".Fragments.AccountDetailsFragment"
+            tools:context=".Features.Login.Views.AccountDetailsFragment"
             android:background="?attr/backgroundColor">
 
         <TextView
