+++ app/src/main/res/layout/fragment_admin.xml
@@ -6,7 +6,7 @@
             android:id="@+id/constraintLayout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".Fragments.AdminFragment"
+            tools:context=".Features.Login.Views.AdminFragment"
             android:background="?attr/backgroundColor">
 
         <TextView
