+++ app/src/main/res/layout/fragment_login.xml
@@ -2,7 +2,7 @@
 <layout xmlns:tools="http://schemas.android.com/tools" xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:android="http://schemas.android.com/apk/res/android">
     <androidx.constraintlayout.widget.ConstraintLayout android:layout_width="match_parent"
                                                        android:layout_height="match_parent"
-                                                       tools:context=".Fragments.LoginFragment"
+                                                       tools:context=".Features.Login.Views.LoginFragment"
                                                        android:id="@+id/constraintLayout"
     android:background="?attr/backgroundColor">
