+++ app/src/main/res/navigation/navigation.xml
@@ -3,7 +3,7 @@
             xmlns:app="http://schemas.android.com/apk/res-auto"
             xmlns:tools="http://schemas.android.com/tools" android:id="@+id/navigation"
             app:startDestination="@id/loginFragment">
-    <fragment android:id="@+id/loginFragment" android:name="mtaubert.loginapplication.Fragments.LoginFragment"
+    <fragment android:id="@+id/loginFragment" android:name="mtaubert.loginapplication.Features.Login.Views.LoginFragment"
               android:label="fragment_login" tools:layout="@layout/fragment_login">
         <action android:id="@+id/action_loginFragment_to_signUpFragment" app:destination="@id/signUpFragment"
                 app:popUpToInclusive="true"/>
@@ -11,24 +11,24 @@
                 app:popUpToInclusive="true"/>
         <action android:id="@+id/action_loginFragment_to_adminFragment" app:destination="@id/adminFragment"/>
     </fragment>
-    <fragment android:id="@+id/signUpFragment" android:name="mtaubert.loginapplication.Fragments.SignUpFragment"
+    <fragment android:id="@+id/signUpFragment" android:name="mtaubert.loginapplication.Features.Login.Views.SignUpFragment"
               android:label="fragment_sign_up" tools:layout="@layout/fragment_sign_up">
         <action android:id="@+id/action_signUpFragment_to_loginFragment" app:destination="@id/loginFragment"
         />
         <action android:id="@+id/action_signUpFragment_to_accountFragment" app:destination="@id/accountFragment"
                 app:popUpTo="@+id/loginFragment" app:popUpToInclusive="true"/>
     </fragment>
-    <fragment android:id="@+id/accountFragment" android:name="mtaubert.loginapplication.Fragments.AccountFragment"
+    <fragment android:id="@+id/accountFragment" android:name="mtaubert.loginapplication.Features.Login.Views.AccountFragment"
               android:label="AccountFragment">
         <action android:id="@+id/action_accountFragment_to_loginFragment" app:destination="@id/loginFragment"/>
         <action
             android:id="@+id/action_accountFragment_to_accountDetailsFragment"
             app:destination="@+id/accountDetailsFragment" />
     </fragment>
-    <fragment android:id="@+id/adminFragment" android:name="mtaubert.loginapplication.Fragments.AdminFragment"
+    <fragment android:id="@+id/adminFragment" android:name="mtaubert.loginapplication.Features.Login.Views.AdminFragment"
               android:label="AdminFragment"/>
     <fragment
         android:id="@+id/accountDetailsFragment"
-        android:name="mtaubert.loginapplication.Fragments.AccountDetailsFragment"
+        android:name="mtaubert.loginapplication.Features.Login.Views.AccountDetailsFragment"
         android:label="AccountDetailsFragment" />
 </navigation>
